celeba-resnet50-attacks/requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
numpy
pandas
Pillow
tqdm
pyyaml
scikit-learn
matplotlib
streamlit
